Organise code into directories
Fixes #74
