+++ UIP.py
@@ -1,4 +1,4 @@
-from scheduler import scheduler
+from uiplib.scheduler import scheduler
 
 if __name__ == "__main__":
     print("Hey this is UIP! you can use it to download"
