+++ uiplib/constants.py
@@ -1,0 +1,7 @@
+import os
+
+WEBSITE = "https://www.reddit.com/r/wallpapers"
+PICS_FOLDER = 'pics'
+NUMBER_OF_IMAGES_TO_PARSE = 24
+CURR_DIR = os.getcwd()
+TIMEOUT = 30*60
